# services/api/docs/swagger.yaml
basePath: /api
definitions:
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  ocserv_user.ActivateUserData:
    properties:
      expire_at:
        example: "2025-12-31"
        type: string
    type: object
  ocserv_user.CreateOcservUserData:
    properties:
      config:
        $ref: '#/definitions/models.OcservUserConfig'
      description:
        example: User for testing VPN access
        maxLength: 1024
        type: string
      expire_at:
        example: "2025-12-31"
        type: string
      group:
        type: string
      password:
        maxLength: 32
        minLength: 2
        type: string
      traffic_size:
        description: 10 GiB
        example: 10737418240
        minimum: 0
        type: integer
      traffic_type:
        enum:
        - Free
        - MonthlyTransmit
        - MonthlyReceive
        - TotallyTransmit
        - TotallyReceive
        example: MonthlyTransmit
        type: string
      unlimited:
        default: false
        example: false
        type: boolean
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - config
    - group
    - password
    - traffic_type
    - username
    type: object
  ocserv_user.OcservUsersResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.OcservUser'
        type: array
    required:
    - meta
    type: object
  ocserv_user.OcservUsersSyncResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/user.Ocpasswd'
        type: array
    required:
    - meta
    type: object
  ocserv_user.StatisticsResponse:
    properties:
      statistics:
        items:
          $ref: '#/definitions/models.DailyTraffic'
        type: array
      total_bandwidths:
        $ref: '#/definitions/repository.TotalBandwidths'
    required:
    - statistics
    - total_bandwidths
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /ocserv/users:
    get:
      consumes:
      - application/json
      description: List of Ocserv Users
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: ocserv username q search
        in: query
        minLength: 2
        name: q
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ocserv_user.OcservUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: List of Ocserv Users
      tags:
      - Ocserv(Users)
    post:
      consumes:
      - application/json
      description: Ocserv User creation
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: ocserv user create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.CreateOcservUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User creation
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}:
    delete:
      consumes:
      - application/json
      description: Ocserv User delete
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User delete
      tags:
      - Ocserv(Users)
    get:
      consumes:
      - application/json
      description: Ocserv user detail
      parameters:
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OcservUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv user detail
      tags:
      - Ocserv(Users)
    patch:
      consumes:
      - application/json
      description: Ocserv User update
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: ocserv user update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.UpdateOcservUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User update
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/activate:
    post:
      consumes:
      - application/json
      description: Restore and activate expired Ocserv User accounts
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: list of ocserv users and expire time
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.ActivateUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Restore and activate expired Ocserv User accounts
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/lock:
    post:
      consumes:
      - application/json
      description: Ocserv User locking
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User locking
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/statistics:
    get:
      consumes:
      - application/json
      description: Ocserv User Statistics
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: date_start
        in: query
        name: date_start
        type: string
      - description: date_end
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ocserv_user.StatisticsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User Statistics
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/unlock:
    post:
      consumes:
      - application/json
      description: Ocserv User unlocking
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User unlocking
      tags:
      - Ocserv(Users)
  /ocserv/users/{username}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Disconnect Ocserv User
      tags:
      - Ocserv(Users)
  /ocserv/users/ocpasswd:
    get:
      consumes:
      - application/json
      description: Ocserv Users from ocpasswd file
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ocserv_user.OcservUsersSyncResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Users from ocpasswd file
      tags:
      - Ocserv(Ocpasswd)
  /ocserv/users/ocpasswd/sync:
    post:
      consumes:
      - application/json
      description: Ocserv Users from ocpasswd file to db
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: list of users with config to sync in db
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.SyncOcpasswdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Users from ocpasswd file to db
      tags:
      - Ocserv(Ocpasswd)
...
swagger: "2.0"